requirements/base.txt
Pillow==7.1.2
PyPDF2==1.26.0
PyYAML==5.3.1
Whoosh==2.7.4
bleach==3.1.5
celery==4.4.7
django-activity-stream==0.8.0
django-celery-beat==2.0.0
django-colorful==1.3
django-cors-headers==3.2.1
django-formtools==2.2
django-mathfilters==1.0.0
django-model-utils==4.0.0
django-mptt==0.12.0
django-pure-pagination==0.3.0
django-qsstats-magic==1.1.0
django-solo==1.1.5
django-stronghold==0.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.2
djangorestframework-recursive==0.1.2
drf-yasg==1.17.1
extract-msg==0.23.3
flanker==0.9.11
flex==6.14.1
furl==2.1.0
fusepy==3.0.1
gevent==20.4.0
graphviz==0.14
gunicorn==20.0.4
mock==4.0.2
node-semver==0.8.0
packaging==20.3
pycountry==20.7.3
pycryptodome==3.9.7
python-dateutil==2.8.1
python-magic==0.4.22
python_gnupg==0.4.6
pytz==2020.1
requests==2.25.1
sh==1.14.1
swagger-spec-validator==2.5.0
whitenoise==5.0.1